clinicalvision_backend/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0

# Data Processing
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0

# Medical Imaging
pydicom==2.4.4
opencv-python-headless==4.9.0.80
scikit-image==0.22.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Database Migrations
alembic>=1.13.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
slowapi==0.1.9
python-magic==0.4.27
redis>=4.5.0
